requirements.txt
# transformers
# torch
faiss-cpu
pymongo
flask
PyPDF2<3.0.0
flasgger==0.9.7b2
python-dotenv
openai
numpy
langchain
tiktoken
